Refactor: 도장 찾기 페이지 구조 변경 csr -> ssr
